backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
opencv-python==4.8.1.78
pytesseract==0.3.10
spacy==3.7.2
python-dotenv==1.0.0
Pillow==10.0.1
numpy==1.24.3
gunicorn==21.2.0
